wip: move utils in properties.
